Added basic fitness function

+++ src/main/java/net/lukemcomber/genetics/biology/fitness/FitnessFunction.java
@@ -4,5 +4,5 @@ import net.lukemcomber.genetics.store.metadata.Performance;
 
 import java.util.function.Function;
 
-public interface FitnessFunction extends Function<Performance,Float> {
+public interface FitnessFunction extends Function<Performance,Double> {
 }
